// src/__generated__/types/Certification.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface Certification {
  organizationId?: string;
  organizationName?: string;
  organizationIcon?: string;
  name?: string;
  description?: string;
  image?: string;
  tags?: string[];
  startDate?: string;
  endDate?: string;
  createdAt?: string;
  updatedAt?: string;
  [k: string]: unknown;
}