// src/__generated__/types/CertificationSubject.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface CertificationSubject {
  id: string;
  certificationId?: string;
  certificationName?: string;
  image?: string;
  [k: string]: unknown;
}